Use ``passes`` to authorized ``allocate_buckets``.
If enough valid passes are supplied for the amount of storage requested then
allow the operation to succeed.  Otherwise, fail with an error.
