+++ src/_zkapauthorizer/_storage_server.py
@@ -24,10 +24,6 @@ from __future__ import (
     absolute_import,
 )
 
-from math import (
-    ceil,
-)
-
 import attr
 from attr.validators import (
     provides,
@@ -56,6 +52,8 @@ from .foolscap import (
     RITokenAuthorizedStorageServer,
 )
 from .storage_common import (
+    BYTES_PER_PASS,
+    required_passes,
     allocate_buckets_message,
     add_lease_message,
     renew_lease_message,
@@ -66,6 +64,15 @@ class MorePassesRequired(Exception):
     def __init__(self, valid_count, required_count):
         self.valid_count = valid_count
         self.required_count = required_count
+
+    def __repr__(self):
+        return "MorePassedRequired(valid_count={}, required_count={})".format(
+            self.valid_count,
+            self.required_count,
+        )
+
+    def __str__(self):
+        return repr(self)
 
 
 @implementer_only(RITokenAuthorizedStorageServer, IReferenceable, IRemotelyCallable)
@@ -78,24 +85,8 @@ class ZKAPAuthorizerStorageServer(Referenceable):
     A class which wraps an ``RIStorageServer`` to insert pass validity checks
     before allowing certain functionality.
     """
-    # The number of bytes we're willing to store for a lease period for each
-    # pass submitted.
-    _BYTES_PER_PASS = 128 * 1024
-
     _original = attr.ib(validator=provides(RIStorageServer))
     _signing_key = attr.ib(validator=instance_of(SigningKey))
-
-    def _required_passes(self, stored_bytes):
-        """
-        Calculate the number of passes that are required to store ``stored_bytes``
-        for one lease period.
-
-        :param int stored_bytes: A number of bytes of storage for which to
-            calculate a price in passes.
-
-        :return int: The number of passes.
-        """
-        return int(ceil(stored_bytes / self._BYTES_PER_PASS))
 
     def _is_invalid_pass(self, message, pass_):
         """
@@ -152,11 +143,11 @@ class ZKAPAuthorizerStorageServer(Referenceable):
             allocate_buckets_message(storage_index),
             passes,
         )
-        required_passes = self._required_passes(len(sharenums) * allocated_size)
-        if len(valid_passes) < required_passes:
+        required_pass_count = required_passes(BYTES_PER_PASS, sharenums, allocated_size)
+        if len(valid_passes) < required_pass_count:
             raise MorePassesRequired(
                 len(valid_passes),
-                required_passes,
+                required_pass_count,
             )
 
         return self._original.remote_allocate_buckets(
